+++ src/object/reject.js
@@ -1,0 +1,9 @@
+import rejectKeys from './rejectKeys';
+
+/**
+ * Alias for rejectKeys.
+ *
+ **/
+export default function reject(...args) {
+  return rejectKeys(...args);
+}