+++ src/object/select.js
@@ -1,0 +1,9 @@
+import selectKeys from './selectKeys';
+
+/**
+ * Alias for selectKeys.
+ *
+ **/
+export default function select(...args) {
+  return selectKeys(...args);
+}