+++ test/tests/object.js
@@ -545,46 +545,46 @@ namespace('Object', function () {
 
   });
 
-  describeInstance('filterValues', function(filterValues) {
+  describeInstance('selectValues', function(selectValues) {
 
     it('should match by primitive matchers', function() {
-      assertObjectEqual(filterValues({a:'a',b:'b'}, 'a'), {a:'a'});
-      assertObjectEqual(filterValues({a:'a',b:'b'}, 'c'), {});
-      assertObjectEqual(filterValues({a:1,b:2}, 2), {b:2});
-      assertObjectEqual(filterValues({a:1,b:2}, 3), {});
-      assertObjectEqual(filterValues({a:true,b:false}, true), {a:true});
-      assertObjectEqual(filterValues({a:true,b:true}, false), {});
+      assertObjectEqual(selectValues({a:'a',b:'b'}, 'a'), {a:'a'});
+      assertObjectEqual(selectValues({a:'a',b:'b'}, 'c'), {});
+      assertObjectEqual(selectValues({a:1,b:2}, 2), {b:2});
+      assertObjectEqual(selectValues({a:1,b:2}, 3), {});
+      assertObjectEqual(selectValues({a:true,b:false}, true), {a:true});
+      assertObjectEqual(selectValues({a:true,b:true}, false), {});
     });
 
     it('should match by regex', function() {
-      assertObjectEqual(filterValues({a:'a',b:'b'}, /[a-c]/), {a:'a',b:'b'});
-      assertObjectEqual(filterValues({a:'a',b:'b'}, /[c-z]/), {});
+      assertObjectEqual(selectValues({a:'a',b:'b'}, /[a-c]/), {a:'a',b:'b'});
+      assertObjectEqual(selectValues({a:'a',b:'b'}, /[c-z]/), {});
     });
 
     it('should match by date', function() {
       var d1 = new Date(2020, 7, 28);
       var d2 = new Date(2020, 7, 29);
-      assertObjectEqual(filterValues({a:d1,b:d2}, new Date(2020, 7, 28)), {a:d1});
-      assertObjectEqual(filterValues({a:d1,b:d1}, new Date(2020, 7, 29)), {});
+      assertObjectEqual(selectValues({a:d1,b:d2}, new Date(2020, 7, 28)), {a:d1});
+      assertObjectEqual(selectValues({a:d1,b:d1}, new Date(2020, 7, 29)), {});
     });
 
     it('should match by function', function() {
-      assertObjectEqual(filterValues({a:1,b:2}, (key, n) => n % 2 === 0), {b:2});
-      assertObjectEqual(filterValues({a:1,b:3}, (key, n) => n % 2 === 0), {});
-      assertObjectEqual(filterValues({a:1,b:2}, (key, n) => n > 5), {});
-      assertObjectEqual(filterValues({a:1,b:2}, (key, n) => n > 0), {a:1,b:2});
+      assertObjectEqual(selectValues({a:1,b:2}, (key, n) => n % 2 === 0), {b:2});
+      assertObjectEqual(selectValues({a:1,b:3}, (key, n) => n % 2 === 0), {});
+      assertObjectEqual(selectValues({a:1,b:2}, (key, n) => n > 5), {});
+      assertObjectEqual(selectValues({a:1,b:2}, (key, n) => n > 0), {a:1,b:2});
     });
 
     it('should match by function when strictly equal', function() {
       var fn1 = function(){};
       var fn2 = function(){};
-      assertObjectEqual(filterValues({a:fn1, b:fn2}, fn2), {b:fn2});
-      assertObjectEqual(filterValues({a:fn1, b:fn1}, fn2), {});
+      assertObjectEqual(selectValues({a:fn1, b:fn2}, fn2), {b:fn2});
+      assertObjectEqual(selectValues({a:fn1, b:fn1}, fn2), {});
     });
 
     it('should match by fuzzy matching', function() {
       assertObjectEqual(
-        filterValues({
+        selectValues({
           1:{ name:'Frank'},
           2:{ name:'James'},
         },
@@ -592,13 +592,13 @@ namespace('Object', function () {
         ),
         { 1: { name:'Frank'} },
       );
-      assertObjectEqual(filterValues({
+      assertObjectEqual(selectValues({
         1:{ name:'Frank'},
         2:{ name:'James'},
       }, {name:'Robert'}),
         {},
       );
-      assertObjectEqual(filterValues({
+      assertObjectEqual(selectValues({
         1:{ name:'Frank'},
         2:{ name:'James'},
       }, {name:/^[A-J]/}),
@@ -607,7 +607,7 @@ namespace('Object', function () {
           2: { name:'James'},
         },
       );
-      assertObjectEqual(filterValues({
+      assertObjectEqual(selectValues({
         1:{ name:'Frank'},
         2:{ name:'James'},
       }, {name:/^[K-Z]/}),
@@ -616,7 +616,7 @@ namespace('Object', function () {
     });
 
     it('should pass correct params', function() {
-      filterValues({a:1}, function (key, val, obj) {
+      selectValues({a:1}, function (key, val, obj) {
         assertEqual(key, 'a');
         assertEqual(val, 1);
         assertObjectEqual(obj, {a:1});
@@ -625,60 +625,60 @@ namespace('Object', function () {
 
     it('should not modify the object', function() {
       const obj = {};
-      assertEqual(obj === filterValues(obj, 1), false);
+      assertEqual(obj === selectValues(obj, 1), false);
     });
 
     it('should handle irregular input', function() {
-      assertObjectEqual(filterValues({a:1}, null), {});
-      assertObjectEqual(filterValues({a:1}, NaN), {});
-      assertError(function() { filterValues({}); });
-      assertError(function() { filterValues(null); });
-      assertError(function() { filterValues('a'); });
-      assertError(function() { filterValues(1); });
+      assertObjectEqual(selectValues({a:1}, null), {});
+      assertObjectEqual(selectValues({a:1}, NaN), {});
+      assertError(function() { selectValues({}); });
+      assertError(function() { selectValues(null); });
+      assertError(function() { selectValues('a'); });
+      assertError(function() { selectValues(1); });
     });
 
   });
 
-  describeInstance('excludeValues', function(excludeValues) {
+  describeInstance('rejectValues', function(rejectValues) {
 
     it('should match by primitive matchers', function() {
-      assertObjectEqual(excludeValues({a:'a',b:'b'}, 'a'), {b:'b'});
-      assertObjectEqual(excludeValues({a:'a',b:'b'}, 'c'), {a:'a',b:'b'});
-      assertObjectEqual(excludeValues({a:1,b:2}, 2), {a:1});
-      assertObjectEqual(excludeValues({a:1,b:2}, 3), {a:1,b:2});
-      assertObjectEqual(excludeValues({a:true,b:false}, true), {b:false});
-      assertObjectEqual(excludeValues({a:true,b:true}, false), {a:true,b:true});
+      assertObjectEqual(rejectValues({a:'a',b:'b'}, 'a'), {b:'b'});
+      assertObjectEqual(rejectValues({a:'a',b:'b'}, 'c'), {a:'a',b:'b'});
+      assertObjectEqual(rejectValues({a:1,b:2}, 2), {a:1});
+      assertObjectEqual(rejectValues({a:1,b:2}, 3), {a:1,b:2});
+      assertObjectEqual(rejectValues({a:true,b:false}, true), {b:false});
+      assertObjectEqual(rejectValues({a:true,b:true}, false), {a:true,b:true});
     });
 
     it('should match by regex', function() {
-      assertObjectEqual(excludeValues({a:'a',b:'b'}, /[a-c]/), {});
-      assertObjectEqual(excludeValues({a:'a',b:'b'}, /[c-z]/), {a:'a',b:'b'});
+      assertObjectEqual(rejectValues({a:'a',b:'b'}, /[a-c]/), {});
+      assertObjectEqual(rejectValues({a:'a',b:'b'}, /[c-z]/), {a:'a',b:'b'});
     });
 
     it('should match by date', function() {
       var d1 = new Date(2020, 7, 28);
       var d2 = new Date(2020, 7, 29);
-      assertObjectEqual(excludeValues({a:d1,b:d2}, new Date(2020, 7, 28)), {b:d2});
-      assertObjectEqual(excludeValues({a:d1,b:d1}, new Date(2020, 7, 29)), {a:d1,b:d1});
+      assertObjectEqual(rejectValues({a:d1,b:d2}, new Date(2020, 7, 28)), {b:d2});
+      assertObjectEqual(rejectValues({a:d1,b:d1}, new Date(2020, 7, 29)), {a:d1,b:d1});
     });
 
     it('should match by function', function() {
-      assertObjectEqual(excludeValues({a:1,b:2}, (key, n) => n % 2 === 0), {a:1});
-      assertObjectEqual(excludeValues({a:1,b:3}, (key, n) => n % 2 === 0), {a:1,b:3});
-      assertObjectEqual(excludeValues({a:1,b:2}, (key, n) => n > 5), {a:1,b:2});
-      assertObjectEqual(excludeValues({a:1,b:2}, (key, n) => n > 0), {});
+      assertObjectEqual(rejectValues({a:1,b:2}, (key, n) => n % 2 === 0), {a:1});
+      assertObjectEqual(rejectValues({a:1,b:3}, (key, n) => n % 2 === 0), {a:1,b:3});
+      assertObjectEqual(rejectValues({a:1,b:2}, (key, n) => n > 5), {a:1,b:2});
+      assertObjectEqual(rejectValues({a:1,b:2}, (key, n) => n > 0), {});
     });
 
     it('should match by function when strictly equal', function() {
       var fn1 = function(){};
       var fn2 = function(){};
-      assertObjectEqual(excludeValues({a:fn1, b:fn2}, fn2), {a:fn1});
-      assertObjectEqual(excludeValues({a:fn1, b:fn1}, fn2), {a:fn1,b:fn1});
+      assertObjectEqual(rejectValues({a:fn1, b:fn2}, fn2), {a:fn1});
+      assertObjectEqual(rejectValues({a:fn1, b:fn1}, fn2), {a:fn1,b:fn1});
     });
 
     it('should match by fuzzy matching', function() {
       assertObjectEqual(
-        excludeValues({
+        rejectValues({
           1:{ name:'Frank'},
           2:{ name:'James'},
         },
@@ -686,7 +686,7 @@ namespace('Object', function () {
         ),
         { 2: { name:'James'} },
       );
-      assertObjectEqual(excludeValues({
+      assertObjectEqual(rejectValues({
         1:{ name:'Frank'},
         2:{ name:'James'},
       }, {name:'Robert'}),
@@ -695,13 +695,13 @@ namespace('Object', function () {
           2:{ name:'James'},
         },
       );
-      assertObjectEqual(excludeValues({
+      assertObjectEqual(rejectValues({
         1:{ name:'Frank'},
         2:{ name:'James'},
       }, {name:/^[A-J]/}),
         {},
       );
-      assertObjectEqual(excludeValues({
+      assertObjectEqual(rejectValues({
         1:{ name:'Frank'},
         2:{ name:'James'},
       }, {name:/^[K-Z]/}),
@@ -713,7 +713,7 @@ namespace('Object', function () {
     });
 
     it('should pass correct params', function() {
-      excludeValues({a:1}, function (key, val, obj) {
+      rejectValues({a:1}, function (key, val, obj) {
         assertEqual(key, 'a');
         assertEqual(val, 1);
         assertObjectEqual(obj, {a:1});
@@ -722,16 +722,16 @@ namespace('Object', function () {
 
     it('should not modify the object', function() {
       const obj = {};
-      assertEqual(obj === excludeValues(obj, 1), false);
+      assertEqual(obj === rejectValues(obj, 1), false);
     });
 
     it('should handle irregular input', function() {
-      assertObjectEqual(excludeValues({a:1}, null), {a:1});
-      assertObjectEqual(excludeValues({a:1}, NaN), {a:1});
-      assertError(function() { excludeValues({}); });
-      assertError(function() { excludeValues(null); });
-      assertError(function() { excludeValues('a'); });
-      assertError(function() { excludeValues(1); });
+      assertObjectEqual(rejectValues({a:1}, null), {a:1});
+      assertObjectEqual(rejectValues({a:1}, NaN), {a:1});
+      assertError(function() { rejectValues({}); });
+      assertError(function() { rejectValues(null); });
+      assertError(function() { rejectValues('a'); });
+      assertError(function() { rejectValues(1); });
     });
 
   });
@@ -835,98 +835,116 @@ namespace('Object', function () {
 
   });
 
-  describeInstance('filter', function(filter) {
+  describeInstance('select,selectKeys', function(selectKeys) {
 
-    it('should filter by enumerated arguments', function() {
-      assertObjectEqual(filter({a:1,b:2}, 'a'), {a:1});
-      assertObjectEqual(filter({a:1,b:2}, 'a', 'b'), {a:1,b:2});
-      assertObjectEqual(filter({a:1,b:2}), {});
-      assertObjectEqual(filter({a:1,b:2}, 'c'), {});
+    it('should select by enumerated arguments', function() {
+      assertObjectEqual(selectKeys({a:1,b:2}, 'a'), {a:1});
+      assertObjectEqual(selectKeys({a:1,b:2}, 'a', 'b'), {a:1,b:2});
+      assertObjectEqual(selectKeys({a:1,b:2}), {});
+      assertObjectEqual(selectKeys({a:1,b:2}, 'c'), {});
     });
 
-    it('should filter by array argument', function() {
-      assertObjectEqual(filter({a:1,b:2}, ['a']), {a:1});
-      assertObjectEqual(filter({a:1,b:2}, ['a', 'b']), {a:1,b:2});
-      assertObjectEqual(filter({a:1,b:2}, []), {});
-      assertObjectEqual(filter({a:1,b:2}, ['c']), {});
+    it('should select by array argument', function() {
+      assertObjectEqual(selectKeys({a:1,b:2}, ['a']), {a:1});
+      assertObjectEqual(selectKeys({a:1,b:2}, ['a', 'b']), {a:1,b:2});
+      assertObjectEqual(selectKeys({a:1,b:2}, []), {});
+      assertObjectEqual(selectKeys({a:1,b:2}, ['c']), {});
     });
 
-    it('should not modify the object', function() {
-      const obj1 = {a:1};
-      const obj2 = filter(obj1, 1);
-      assertFalse(obj1 === obj2);
-    });
-
-    it('should handle irregular input', function() {
-      assertObjectEqual(filter({a:1}, null), {});
-      assertObjectEqual(filter({a:1}, NaN), {});
-      assertError(function() { filter(null); });
-      assertError(function() { filter('a'); });
-      assertError(function() { filter(1); });
-    });
-
-  });
-
-  describeInstance('exclude', function(exclude) {
-
-    it('should exclude by enumerated arguments', function() {
-      assertObjectEqual(exclude({a:1,b:2}, 'a'), {b:2});
-      assertObjectEqual(exclude({a:1,b:2}, 'a', 'b'), {});
-      assertObjectEqual(exclude({a:1,b:2}), {a:1,b:2});
-      assertObjectEqual(exclude({a:1,b:2}, 'c'), {a:1,b:2});
-    });
-
-    it('should exclude by array argument', function() {
-      assertObjectEqual(exclude({a:1,b:2}, ['a']), {b:2});
-      assertObjectEqual(exclude({a:1,b:2}, ['a', 'b']), {});
-      assertObjectEqual(exclude({a:1,b:2}, []), {a:1,b:2});
-      assertObjectEqual(exclude({a:1,b:2}, ['c']), {a:1,b:2});
+    it('should select by regex', function() {
+      assertObjectEqual(selectKeys({a:1,b:2}, /a/), {a:1});
+      assertObjectEqual(selectKeys({a:1,b:2}, /[ab]/), {a:1,b:2});
+      assertObjectEqual(selectKeys({a:1,b:2}, /c/), {});
     });
 
     it('should not modify the object', function() {
       const obj1 = {a:1};
-      const obj2 = exclude(obj1, 1);
+      const obj2 = selectKeys(obj1, 1);
       assertFalse(obj1 === obj2);
     });
 
     it('should handle irregular input', function() {
-      assertObjectEqual(exclude({a:1}, null), {a:1});
-      assertObjectEqual(exclude({a:1}, NaN), {a:1});
-      assertError(function() { exclude(null); });
-      assertError(function() { exclude('a'); });
-      assertError(function() { exclude(1); });
+      assertObjectEqual(selectKeys({a:1}, null), {});
+      assertObjectEqual(selectKeys({a:1}, NaN), {});
+      assertError(function() { selectKeys(null); });
+      assertError(function() { selectKeys('a'); });
+      assertError(function() { selectKeys(1); });
     });
 
   });
 
-  describeInstance('remove', function(remove) {
+  describeInstance('reject,rejectKeys', function(rejectKeys) {
+
+    it('should reject by enumerated arguments', function() {
+      assertObjectEqual(rejectKeys({a:1,b:2}, 'a'), {b:2});
+      assertObjectEqual(rejectKeys({a:1,b:2}, 'a', 'b'), {});
+      assertObjectEqual(rejectKeys({a:1,b:2}), {a:1,b:2});
+      assertObjectEqual(rejectKeys({a:1,b:2}, 'c'), {a:1,b:2});
+    });
+
+    it('should reject by array argument', function() {
+      assertObjectEqual(rejectKeys({a:1,b:2}, ['a']), {b:2});
+      assertObjectEqual(rejectKeys({a:1,b:2}, ['a', 'b']), {});
+      assertObjectEqual(rejectKeys({a:1,b:2}, []), {a:1,b:2});
+      assertObjectEqual(rejectKeys({a:1,b:2}, ['c']), {a:1,b:2});
+    });
+
+    it('should reject by regex', function() {
+      assertObjectEqual(rejectKeys({a:1,b:2}, /a/), {b:2});
+      assertObjectEqual(rejectKeys({a:1,b:2}, /[ab]/), {});
+      assertObjectEqual(rejectKeys({a:1,b:2}, /c/), {a:1,b:2});
+    });
+
+    it('should not modify the object', function() {
+      const obj1 = {a:1};
+      const obj2 = rejectKeys(obj1, 1);
+      assertFalse(obj1 === obj2);
+    });
+
+    it('should handle irregular input', function() {
+      assertObjectEqual(rejectKeys({a:1}, null), {a:1});
+      assertObjectEqual(rejectKeys({a:1}, NaN), {a:1});
+      assertError(function() { rejectKeys(null); });
+      assertError(function() { rejectKeys('a'); });
+      assertError(function() { rejectKeys(1); });
+    });
+
+  });
+
+  describeInstance('remove,removeKeys', function(removeKeys) {
 
     it('should remove by enumerated arguments', function() {
-      assertObjectEqual(remove({a:1,b:2}, 'a'), {b:2});
-      assertObjectEqual(remove({a:1,b:2}, 'a', 'b'), {});
-      assertObjectEqual(remove({a:1,b:2}), {a:1,b:2});
-      assertObjectEqual(remove({a:1,b:2}, 'c'), {a:1,b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, 'a'), {b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, 'a', 'b'), {});
+      assertObjectEqual(removeKeys({a:1,b:2}), {a:1,b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, 'c'), {a:1,b:2});
     });
 
     it('should remove by array argument', function() {
-      assertObjectEqual(remove({a:1,b:2}, ['a']), {b:2});
-      assertObjectEqual(remove({a:1,b:2}, ['a', 'b']), {});
-      assertObjectEqual(remove({a:1,b:2}, []), {a:1,b:2});
-      assertObjectEqual(remove({a:1,b:2}, ['c']), {a:1,b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, ['a']), {b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, ['a', 'b']), {});
+      assertObjectEqual(removeKeys({a:1,b:2}, []), {a:1,b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, ['c']), {a:1,b:2});
+    });
+
+    it('should remove by regex', function() {
+      assertObjectEqual(removeKeys({a:1,b:2}, /a/), {b:2});
+      assertObjectEqual(removeKeys({a:1,b:2}, /[ab]/), {});
+      assertObjectEqual(removeKeys({a:1,b:2}, /c/), {a:1,b:2});
     });
 
     it('should modify the object', function() {
       const obj1 = {a:1};
-      const obj2 = remove(obj1, 1);
+      const obj2 = removeKeys(obj1, 1);
       assertTrue(obj1 === obj2);
     });
 
     it('should handle irregular input', function() {
-      assertObjectEqual(remove({a:1}, null), {a:1});
-      assertObjectEqual(remove({a:1}, NaN), {a:1});
-      assertError(function() { remove(null); });
-      assertError(function() { remove('a'); });
-      assertError(function() { remove(1); });
+      assertObjectEqual(removeKeys({a:1}, null), {a:1});
+      assertObjectEqual(removeKeys({a:1}, NaN), {a:1});
+      assertError(function() { removeKeys(null); });
+      assertError(function() { removeKeys('a'); });
+      assertError(function() { removeKeys(1); });
     });
 
   });